Make PlaywrightHelper only have static methods
Do not extend a super class for
playwright tests
