+++ vaadin-testbench-loadtest/load-tests/demo-web-app-playwright/src/test/java/com/vaadin/laboratory/views/scenario/HelloWorldPlaywrightIT.java
@@ -8,11 +8,17 @@
  */
 package com.vaadin.laboratory.views.scenario;
 
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import com.microsoft.playwright.Browser;
+import com.microsoft.playwright.BrowserContext;
+import com.microsoft.playwright.BrowserType;
 import com.microsoft.playwright.Page;
+import com.microsoft.playwright.Playwright;
 import com.microsoft.playwright.options.AriaRole;
-import com.vaadin.testbench.loadtest.AbstractPlaywrightHelper;
+import com.vaadin.testbench.loadtest.PlaywrightHelper;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 
@@ -25,15 +31,41 @@ import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertTha
  * <li>Enter a name and press the "Say hello" button</li>
  * </ol>
  */
-public class HelloWorldPlaywrightIT extends AbstractPlaywrightHelper {
+public class HelloWorldPlaywrightIT {
 
     private static final String TEST_NAME = "Vaadin User";
 
+    private Playwright playwright;
+    private Browser browser;
+    private BrowserContext context;
+    private Page page;
+
+    @BeforeEach
+    public void setUp() {
+        playwright = Playwright.create();
+        browser = playwright.chromium()
+                .launch(new BrowserType.LaunchOptions().setHeadless(true));
+        context = PlaywrightHelper.createBrowserContext(browser);
+        page = context.newPage();
+        page.navigate(PlaywrightHelper.getBaseUrl() + "/");
+    }
+
+    @AfterEach
+    public void tearDown() {
+        if (context != null) {
+            context.close();
+        }
+        if (browser != null) {
+            browser.close();
+        }
+        if (playwright != null) {
+            playwright.close();
+        }
+    }
+
     @Test
     public void helloWorldWorkflow() {
-        // 1. open the view - done automatically by AbstractPlaywrightHelper
-
-        // 2. enter a name and press the say hello button
+        // Enter a name and press the say hello button
         enterNameAndSayHello();
     }
 
@@ -48,10 +80,5 @@ public class HelloWorldPlaywrightIT extends AbstractPlaywrightHelper {
         // Verify notification
         assertThat(page.locator("vaadin-notification-card"))
                 .containsText("Hello " + TEST_NAME);
-    }
-
-    @Override
-    public String getViewName() {
-        return "";
     }
 }
